Moving IntegralGrader out of plugins
